requirements.txt
aniso8601==9.0.1
apispec==5.2.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.12.3
binance-pay-connector==1.0.0
blinker==1.4
boto3==1.24.19
botocore==1.27.19
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
click==8.1.3
coinbase-commerce==1.0.1
cryptography==37.0.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask==2.1.2
geopy==2.4.1
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googletrans==2.4.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.3
iniconfig==1.1.1
ipython==8.18.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.6.1
langdetect==1.0.9
MarkupSafe==2.1.1
marshmallow==3.17.0
mysqlclient==2.1.1
nltk==3.8.1
oauthlib==3.2.2
openai==1.11.1
packaging==21.3
pandas==1.1.5
pdfkit==1.0.0
pillow==10.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pylogbeat==2.0.0
PyMySQL==1.0.2
pyparsing==3.0.9
pySqsListener==0.8.10
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis-om==0.0.27
redis==4.3.4
redisbloom==0.4.0
redisearch==2.0.0
redisgraph==2.3.0
redistimeseries==1.4.3
rejson==0.5.4
requests-oauthlib==1.3.1
requests==2.28.1
s3transfer==0.6.0
selenium==4.3.0
sentry-sdk==1.6.0
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.39
stripe==8.0.0
toml==0.10.2
twilio==7.9.3
uritemplate==4.1.1
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.0.3
gspread